QInchworm.jl

QInchworm.QInchworm

QInchworm.SectorBlockMatrix

QInchworm.operator_to_sector_block_matrix

QInchworm.QInchworm — Module

A quasi Monte Carlo inchworm impurity solver for multi-orbital fermionic models.

QInchworm.SectorBlockMatrix — Type

Complex block matrix stored as a dictionary of non-vanishing blocks.

An element of the dictionary has the form right block index => (left block index, block).

Base.zeros — Method

zeros(
    _::Type{Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}},
    ed::KeldyshED.EDCore
) -> Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}

Construct a block-diagonal complex matrix, whose block structure is consistent with the invariant subspace partition of a given KeldyshED.EDCore object. All matrix elements of the stored blocks are set to zero.

QInchworm.operator_to_sector_block_matrix — Method

operator_to_sector_block_matrix(
    ed::KeldyshED.EDCore,
    op::KeldyshED.Operators.OperatorExpr
) -> Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}

Returns the SectorBlockMatrix representation of the many-body operator.

Sub Modules

QInchworm.qmc_integrate

QInchworm.configuration

QInchworm.ppgf

QInchworm.diagrammatics