QInchworm.configuration

The configuration module is a framework for representing generic pseudo particle diagrams with a back-bone of pseudo particle propagators and insertions of pseudo particle interactions at fixed times.

There is also functionality to represent Inch-Worm diagrams with a fixed "inch" time where the back bone propagator switches from bold to bare.

The building blocks common for all kinds of diagrams are contained in the Expansion struct. Given an Expansion a particular diagram with fixed times can be represented by the Configuration struct.

Given an Expansion and a Configuration the eval function can be used to evaluate the value of the "configuration"/diagram.

Pseudo particle interactions

In pseudo particle strong coupling expansions the notion "hybridization insertions" can be seen as a retarded interaction between the pseudo particles.

One such pseudo particle interaction is defined by a scalar hybridization function Green's function and an initial and final many-body operator.

Thus, a pseudo particle Expansion is defined, in part, by the list of all possible pseudo particle interactions in the system, where each pseudo particle interaction is represented using the struct InteractionPair.

Pseudo-particle symmetry sectors

The pseudo-particle propagator and the many-body operators are represented as block matrices in the local Hilbert space. The blocking is performed by KeldyshED.jl based on the symmetries of the local Hamiltonian.

To represent many-body operators the SectorBlockMatrix type is used.

Module Index

QInchworm.configuration.Configuration

QInchworm.configuration.InteractionEnum

QInchworm.configuration.Node

QInchworm.configuration.Node

QInchworm.configuration.NodePair

QInchworm.configuration.OperatorReference

QInchworm.configuration.InchNode

QInchworm.configuration.Nodes

QInchworm.configuration.OperatorNode

QInchworm.configuration.eval

QInchworm.configuration.get_block_matrix

QInchworm.configuration.is_inch_node

QInchworm.configuration.is_operator_node

QInchworm.configuration.sector_block_matrix_from_ppgf

Module API Documentation

QInchworm.configuration.Configuration — Type

struct Configuration

The Configuration struct defines a single diagram in a peudo-particle expansion with fixed insertions of pseduo-particle interactions and auxilliary operators.

nodes::Vector{QInchworm.configuration.Node}: List of nodes in time with associated operators

pairs::Vector{QInchworm.configuration.NodePair}: List of pairs of nodes in time associated with a hybridization propagator

parity::Float64: Parity of the diagram p = (-1)^N, with N number of hybridization line crossings

determinants::Vector{QInchworm.configuration.Determinant}: List of groups of time nodes associated with expansion determinants

paths::Vector{Vector{Tuple{Int64, Int64}}}: List of precomputed trace paths

split_node_idx::Union{Nothing, Int64}: Position of the node that splits the integration domain into two simplices

op_node_idx::Union{Nothing, Tuple{Int64, Int64}}: Positions of two operator nodes used to measure correlation functions

pair_node_idxs::Vector{Int64}: Positions of nodes coupled by pair interactions

QInchworm.configuration.InteractionEnum — Type

Interaction type classification using @enum

Possible values: pair_flag, determinant_flag, identity_flag, inch_flag, operator_flag

primitive type InteractionEnum <: Enum{Int32} 32

QInchworm.configuration.Node — Type

struct Node

Node in time with associated operator

time::Keldysh.BranchPoint: Contour time point

operator_ref::QInchworm.configuration.OperatorReference: Reference to operator

QInchworm.configuration.Node — Method

Node(
    time::Keldysh.BranchPoint
) -> QInchworm.configuration.Node

Returns a node at time time::Time with an associated identity operator.

QInchworm.configuration.NodePair — Type

struct NodePair

Node with pair of times and an associated interaction index.

time_f::Keldysh.BranchPoint: Final time

time_i::Keldysh.BranchPoint: Initial time

index::Int64: Index for interaction

QInchworm.configuration.OperatorReference — Type

struct OperatorReference

Lightweight reference to an Expansion operator.

kind::QInchworm.configuration.InteractionEnum: Interaction type of operator

interaction_index::Int64: Index for interaction

operator_index::Int64: Index to operator

QInchworm.configuration.InchNode — Method

InchNode(
    time::Keldysh.BranchPoint
) -> QInchworm.configuration.Node

Returns an "inch" node at time time::Time with an associated identity operator.

The Inch node triggers the configuration evaluator to switch from bold to bare pseudo particle propagator.

QInchworm.configuration.Nodes — Method

Nodes(
    pair::QInchworm.configuration.NodePair
) -> Vector{QInchworm.configuration.Node}

Returns a list of Node's corresponding to the pair pair::NodePair.

QInchworm.configuration.OperatorNode — Method

OperatorNode(
    time::Keldysh.BranchPoint,
    interaction_index::Int64,
    operator_index::Int64
) -> QInchworm.configuration.Node

Returns an operator node at time time::Time with an associated operator.

QInchworm.configuration.eval — Method

eval(
    exp::QInchworm.expansion.Expansion,
    conf::QInchworm.configuration.Configuration
) -> Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}

Evaluate the configuration conf in the pseud-particle expansion exp.

QInchworm.configuration.get_block_matrix — Method

get_block_matrix(
    exp::QInchworm.expansion.Expansion,
    node::QInchworm.configuration.Node
) -> Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}

Returns the SectorBlockMatrix representation of the many-body operator at the given `node::Node'.

QInchworm.configuration.is_inch_node — Method

is_inch_node(node::QInchworm.configuration.Node) -> Bool

Returns true if the node is an "inch" node.

QInchworm.configuration.is_operator_node — Method

is_operator_node(node::QInchworm.configuration.Node) -> Bool

Returns true if the node is an operator node.

QInchworm.configuration.sector_block_matrix_from_ppgf — Method

sector_block_matrix_from_ppgf(
    z2::Keldysh.BranchPoint,
    z1::Keldysh.BranchPoint,
    P::Array{MatrixGF<:Keldysh.AbstractTimeGF{ComplexF64, false}, 1}
) -> Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}

Returns the SectorBlockMatrix representation of the pseudo particle propagator evaluated at the times z1 and z2.