QInchworm.diagrammatics

QInchworm.diagrammatics.Diagram

QInchworm.diagrammatics.Topology

QInchworm.diagrammatics.count_doubly_k_connected

QInchworm.diagrammatics.generate_topologies

QInchworm.diagrammatics.generate_topologies_impl

QInchworm.diagrammatics.is_doubly_k_connected

QInchworm.diagrammatics.is_doubly_k_connected

QInchworm.diagrammatics.is_k_connected

QInchworm.diagrammatics.iscrossing

QInchworm.diagrammatics.n_crossings

QInchworm.diagrammatics.pair_partitions

QInchworm.diagrammatics.pair_partitions

QInchworm.diagrammatics.pair_partitions

QInchworm.diagrammatics.parity

QInchworm.diagrammatics.pop_pair

QInchworm.diagrammatics.split_doubly_k_connected

QInchworm.diagrammatics.split_k_connected

QInchworm.diagrammatics.traverse_crossing_graph_dfs!

QInchworm.diagrammatics.Diagram — Type

Diagram with a topology and tuple of pseudo particle interaction pair indices

QInchworm.diagrammatics.Topology — Type

struct Topology

Datatype for diagram topology. A topology of order $n$ consists of a partition of the ordered set $s = \{1,...,2n\}$ into $n$ pairs \{(x(1), x(2)), ..., (x(2n-1), x(2n))\} where $x$ is a permutation of $s$. Diagrammatically a topology can be thought of as a set of arcs connecting vertices located at $\{1,...,2n\}$. The parity of the topology is the sign of the permutation $x$.

order::Int64

pairs::Vector{Pair{Int64, Int64}}

parity::Int64

QInchworm.diagrammatics.count_doubly_k_connected — Method

count_doubly_k_connected(
    pairs::Vector{Pair{Int64, Int64}},
    k::Int64
) -> Int64

Given a vector of pairs, count the doubly k-connected ones.

QInchworm.diagrammatics.generate_topologies — Method

generate_topologies(
    n::Int64
) -> Vector{QInchworm.diagrammatics.Topology}

Return topologies of order $n$, efficiently computing the permutation sign for each.

QInchworm.diagrammatics.generate_topologies_impl — Method

generate_topologies_impl(
    topology_partial::QInchworm.diagrammatics.Topology,
    unpaired::Vector{Int64}
) -> Vector{QInchworm.diagrammatics.Topology}

Given a partial topology and a vector of unpaired vertices, return a vector of complete topologies, efficiently computing the permutation sign.

QInchworm.diagrammatics.is_doubly_k_connected — Method

is_doubly_k_connected(
    p::Pair{Int64, Int64},
    k::Int64
) -> Bool

Returns true if a given pair has one index <= k and the other index > k.

QInchworm.diagrammatics.is_doubly_k_connected — Method

is_doubly_k_connected(
    t::QInchworm.diagrammatics.Topology,
    k::Int64
) -> Bool

Given a topology, check if every connected component of the graph induced by crossings between the arcs contains a pair with one index <= k and the other index > k

QInchworm.diagrammatics.is_k_connected — Method

is_k_connected(
    t::QInchworm.diagrammatics.Topology,
    k::Int64
) -> Bool

Given a topology, check if every connected component of the graph induced by crossings between the arcs contains a pair with index <= k

QInchworm.diagrammatics.iscrossing — Method

iscrossing(
    p1::Pair{Int64, Int64},
    p2::Pair{Int64, Int64}
) -> Bool

Returns true if two arcs cross

Let $p_1 = (a, b)$, $p_2 = (x, y)$ represent two arcs, where without loss of generality we assume $a < b$ and $x < y$. Now consider the order the points $\{a, b, x, y\}$. The orderings $abxy$, $axyb$, $xyab$ are all non-crossing while $axby$ and $xayb$ are crossing.

QInchworm.diagrammatics.n_crossings — Method

n_crossings(top::QInchworm.diagrammatics.Topology) -> Int64

Returns the number of crossing arcs in a topology.

QInchworm.diagrammatics.pair_partitions — Method

pair_partitions(
    n::Int64
) -> Vector{Vector{Pair{Int64, Int64}}}

Return all partitions of $\{1,...,2n\}$ into $n$ pairs.

QInchworm.diagrammatics.pair_partitions — Method

pair_partitions(
    vertices::Vector{Int64}
) -> Vector{Vector{Pair{Int64, Int64}}}

Return all partitions of the vertices into pairs.

QInchworm.diagrammatics.pair_partitions — Method

pair_partitions(
    pairs::Vector{Pair{Int64, Int64}},
    unpaired::Vector{Int64}
) -> Vector{Vector{Pair{Int64, Int64}}}

Given a vector of pairs representing a partial partition of the vertices and a vector of unpaired vertices, return a vector of complete partitions.

QInchworm.diagrammatics.parity — Method

parity(top::QInchworm.diagrammatics.Topology) -> Int64

Returns the parity of the permutation matrix of the topolgy.

QInchworm.diagrammatics.pop_pair — Method

pop_pair(v::Vector, i, j) -> Tuple{Pair, Any}

returns the pair v[i] => v[j] and a copy of the vector v with elements i,j removed

QInchworm.diagrammatics.split_doubly_k_connected — Method

split_doubly_k_connected(
    pairs::Vector{Pair{Int64, Int64}},
    k::Int64
) -> Tuple{Vector{Pair{Int64, Int64}}, Vector{Pair{Int64, Int64}}}

Given a vector of pairs, split it into a 'connected' set containing pairs with one index <= k and the other index > k and a disconnected set containing the rest

QInchworm.diagrammatics.split_k_connected — Method

split_k_connected(
    pairs::Vector{Pair{Int64, Int64}},
    k::Int64
) -> Tuple{Vector{Pair{Int64, Int64}}, Vector{Pair{Int64, Int64}}}

Given a vector of pairs, split it into a 'connected' set containing pairs with an index <= k and a disconnected set containing the rest

QInchworm.diagrammatics.traverse_crossing_graph_dfs! — Method

traverse_crossing_graph_dfs!(
    connected::Vector{Pair{Int64, Int64}},
    disconnected::Vector{Pair{Int64, Int64}}
)

Given a vector of 'connected' arcs and a vector of 'disconnected' arcs recursively add disconnected to connected if they cross with any connected. This is done by traversing the crossing graph using depth first search.