QInchworm.expansion

QInchworm.expansion.AllPPGFSectorTypes — Type

Supported container types for a single block of an atomic propagator (PPGF)

QInchworm.expansion.AllPPGFTypes — Type

Supported container types for a full atomic propagator (PPGF)

QInchworm.expansion.set_bold_ppgf! — Function

set_bold_ppgf!(
    expansion::QInchworm.expansion.Expansion{ScalarGF<:Keldysh.AbstractTimeGF{ComplexF64, true}, Vector{QInchworm.spline_gf.IncSplineImaginaryTimeGF{ComplexF64, false}}},
    τ_i::Keldysh.TimeGridPoint,
    τ_f::Keldysh.TimeGridPoint,
    val::Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}
)

Set the value of expansion.P corresponding to a given pair of imaginary time points $(\tau_f, \tau_i)$. This method is defined for the spline-interpolated imaginary-time propagators.

Parameters

expansion: Pseudo-particle expansion.

τ_i: Initial imaginary time $\tau_i$.

τ_f: Final imaginary time $\tau_f$.

val: Block matrix to set $P(\tau_f, \tau_i)$ to.

set_bold_ppgf!(
    expansion::QInchworm.expansion.Expansion,
    t_i::Keldysh.TimeGridPoint,
    t_f::Keldysh.TimeGridPoint,
    val::Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}
)

Set the value of expansion.P corresponding to a given pair of contour time points $(t_f, t_i)$.

Parameters

expansion: Pseudo-particle expansion.

t_i: Initial imaginary time $t_i$.

t_f: Final imaginary time $t_f$.

val: Block matrix to set $P(t_f, t_i)$ to.

QInchworm.expansion.set_bold_ppgf_at_order! — Function

set_bold_ppgf_at_order!(
    expansion::QInchworm.expansion.Expansion{ScalarGF<:Keldysh.AbstractTimeGF{ComplexF64, true}, Vector{QInchworm.spline_gf.IncSplineImaginaryTimeGF{ComplexF64, false}}},
    order::Integer,
    τ_i::Keldysh.TimeGridPoint,
    τ_f::Keldysh.TimeGridPoint,
    val::Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}
)

Set the value of expansion.P_orders corresponding to a given expansion order and to a pair of imaginary time points $(\tau_f, \tau_i)$. This method is defined for the spline-interpolated imaginary-time propagators.

Parameters

expansion: Pseudo-particle expansion.

order: Expansion order.

τ_i: Initial imaginary time $\tau_i$.

τ_f: Final imaginary time $\tau_f$.

val: Block matrix to set $P(\tau_f, \tau_i)$ to.

set_bold_ppgf_at_order!(
    expansion::QInchworm.expansion.Expansion,
    order::Integer,
    t_i::Keldysh.TimeGridPoint,
    t_f::Keldysh.TimeGridPoint,
    val::Dict{Int64, Tuple{Int64, Matrix{ComplexF64}}}
)

Set the value of expansion.P_orders corresponding to a given expansion order and to a pair of contour time points $(t_f, t_i)$.

Parameters

expansion: Pseudo-particle expansion.

order: Expansion order.

t_i: Initial imaginary time $t_i$.

t_f: Final imaginary time $t_f$.

val: Block matrix to set $P(t_f, t_i)$ to.