QInchworm.qmc_integrate

QInchworm.qmc_integrate — Module

Quasi Monte Carlo integration routines.

Concepts and notation used here are introduced in

Quantum Quasi-Monte Carlo Technique for Many-Body Perturbative Expansions
M. Maček, P. T. Dumitrescu, C. Bertrand, B.Triggs, O. Parcollet, and X. Waintal
Phys. Rev. Lett. 125, 047702 (2020)

Integration domain transformations Sort and Root are defined in

Transforming low-discrepancy sequences from a cube to a simplex
T. Pillards and R. Cools
J. Comput. Appl. Math. 174, 29 (2005)

Model functions and domain transformations

QInchworm.qmc_integrate.make_model_function — Function

make_model_function(
    c::Keldysh.AbstractContour,
    t_f::Keldysh.BranchPoint,
    h::Vector
) -> QInchworm.qmc_integrate.var"#1#3"

Make the model function $p_d(\mathbf{u}) = \prod_{i=1}^d h_i(u_{i-1} - u_i)$ out of a list of functions $h_i(v)$.

$u_0$ is the distance measured along the time contour c to the point t_f.

QInchworm.qmc_integrate.make_exp_model_function — Function

make_exp_model_function(
    c::Keldysh.AbstractContour,
    t_f::Keldysh.BranchPoint,
    τ::Real,
    d::Int64
) -> QInchworm.qmc_integrate.var"#5#6"

Make the model function $p_d(\mathbf{u}) = \prod_{i=1}^d h(u_{i-1} - u_i)$ out of an exponential function $h(v) = e^{-v/\tau}$.

$u_0$ is the distance measured along the time contour c to the point t_f.

QInchworm.qmc_integrate.make_exp_trans — Function

make_exp_trans(
    c::Keldysh.AbstractContour,
    t_f::Keldysh.BranchPoint,
    τ::Real
) -> QInchworm.qmc_integrate.var"#7#8"

Make a transformation from the unit hypercube $\mathbf{x}\in[0, 1]^d$ to a $d$-dimensional simplex $u_f > u_1 > u_2 > \ldots > u_{d-1} > u_d > -\infty$ induced by the model function $p_d(\mathbf{u}) = \prod_{i=1}^d e^{(u_{i-1} - u_i)/\tau}$.

The target variables u_i are reference values of time points measured as distances along the contour c. $u_0$ is the distance to the point t_f.

QInchworm.qmc_integrate.exp_p_norm — Function

exp_p_norm(τ::Real, d::Int64) -> Real

Norm of the model function $p_d(\mathbf{u}) = \prod_{i=1}^d e^{(u_{i-1} - u_i)/\tau}$.

Quasi Monte Carlo integration

QInchworm.qmc_integrate.qmc_integral — Function

qmc_integral(f)
qmc_integral(f, init; p, p_norm, trans, seq, N)

Quasi Monte Carlo integration with warping.

Parameters

f: Integrand.

init: Initial (zero) value of the integral.

p: Positive model function $p_d(\mathbf{u})$.

p_norm: Integral of $p_d(\mathbf{u})$ over the $\mathbf{u}$-domain.

trans: Transformation from $\mathbf{x}\in[0, 1]^d$ onto the $\mathbf{u}$-domain.

seq: Quasi-random sequence generator.

N: The number of points to be taken from the quasi-random sequence.

Returns

Value of the integral.

QInchworm.qmc_integrate.qmc_integral_n_samples — Function

qmc_integral_n_samples(f)
qmc_integral_n_samples(
    f,
    init;
    p,
    p_norm,
    trans,
    seq,
    N_samples
)

Quasi Monte Carlo integration with warping.

This function takes a specified number of valid (non-nothing) samples of the integrand.

Parameters

f: Integrand.

init: Initial (zero) value of the integral.

p: Positive model function $p_d(\mathbf{u})$.

p_norm: Integral of $p_d(\mathbf{u})$ over the $\mathbf{u}$-domain.

trans: Transformation from $\mathbf{x}\in[0, 1]^d$ onto the $\mathbf{u}$-domain.

seq: Quasi-random sequence generator.

N_samples: The number of samples to be taken.

Returns

Value of the integral.

QInchworm.qmc_integrate.qmc_time_ordered_integral — Function

qmc_time_ordered_integral(
    f,
    d::Int64,
    c::Keldysh.AbstractContour,
    t_i::Keldysh.BranchPoint,
    t_f::Keldysh.BranchPoint;
    init,
    seq,
    τ,
    N
)

Evaluate a $d$-dimensional contour-ordered integral of a function $f(\mathbf{t})$,

\[ \int_{t_i}^{t_f} dt_1 \int_{t_i}^{t_1} dt_2 \ldots \int_{t_i}^{t_{d-1}} dt_d \ f(t_1, t_2, \ldots, t_d)\]

using the Sobol sequence for quasi-random sampling and the exponential model function $p_d(\mathbf{t}) = \prod_{i=1}^d e^{(t_{i-1} - t_i)/\tau}$.

Parameters

f: Integrand.

d: Dimensionality of the integral.

c: Time contour to integrate over.

t_i: Starting time point on the contour.

t_f: Final time point on the contour.

init: Initial (zero) value of the integral.

seq: Quasi-random sequence generator.

τ: Decay parameter of the exponential model function.

N: The number of points to be taken from the quasi-random sequence.

Returns

Value of the integral.

QInchworm.qmc_integrate.qmc_time_ordered_integral_n_samples — Function

qmc_time_ordered_integral_n_samples(
    f,
    d::Int64,
    c::Keldysh.AbstractContour,
    t_i::Keldysh.BranchPoint,
    t_f::Keldysh.BranchPoint;
    init,
    seq,
    τ,
    N_samples
)

Evaluate a $d$-dimensional contour-ordered integral of a function $f(\mathbf{t})$,

\[ \int_{t_i}^{t_f} dt_1 \int_{t_i}^{t_1} dt_2 \ldots \int_{t_i}^{t_{d-1}} dt_d \ f(t_1, t_2, \ldots, t_d)\]

using the Sobol sequence for quasi-random sampling and the exponential model function $p_d(\mathbf{t}) = \prod_{i=1}^d e^{(t_{i-1} - t_i)/\tau}$.

This function evaluates the integrand a specified number of times while discarding any transformed sequence points that fall outside the integration domain.

Parameters

f: Integrand.

d: Dimensionality of the integral.

c: Time contour to integrate over.

t_i: Starting time point on the contour.

t_f: Final time point on the contour.

init: Initial (zero) value of the integral.

seq: Quasi-random sequence generator.

τ: Decay parameter of the exponential model function.

N_samples: The number of samples to be taken.

Returns

Value of the integral.

QInchworm.qmc_integrate.qmc_time_ordered_integral_sort — Function

qmc_time_ordered_integral_sort(
    f,
    d::Int64,
    c::Keldysh.AbstractContour,
    t_i::Keldysh.BranchPoint,
    t_f::Keldysh.BranchPoint;
    init,
    seq,
    N
)

Evaluate a $d$-dimensional contour-ordered integral of a function $f(\mathbf{t})$,

\[ \int_{t_i}^{t_f} dt_1 \int_{t_i}^{t_1} dt_2 \ldots \int_{t_i}^{t_{d-1}} dt_d \ f(t_1, t_2, \ldots, t_d)\]

using the Sobol sequence for quasi-random sampling and the 'Sort' transform.

Parameters

f: Integrand.

d: Dimensionality of the integral.

c: Time contour to integrate over.

t_i: Starting time point on the contour.

t_f: Final time point on the contour.

init: Initial (zero) value of the integral.

seq: Quasi-random sequence generator.

N: The number of points to be taken from the quasi-random sequence.

Returns

Value of the integral.

QInchworm.qmc_integrate.qmc_time_ordered_integral_root — Function

qmc_time_ordered_integral_root(
    f,
    d::Int64,
    c::Keldysh.AbstractContour,
    t_i::Keldysh.BranchPoint,
    t_f::Keldysh.BranchPoint;
    init,
    seq,
    N
)

Evaluate a $d$-dimensional contour-ordered integral of a function $f(\mathbf{t})$,

\[ \int_{t_i}^{t_f} dt_1 \int_{t_i}^{t_1} dt_2 \ldots \int_{t_i}^{t_{d-1}} dt_d \ f(t_1, t_2, \ldots, t_d)\]

using the Sobol sequence for quasi-random sampling and the 'Root' transform.

Parameters

f: Integrand.

d: Dimensionality of the integral.

c: Time contour to integrate over.

t_i: Starting time point on the contour.

t_f: Final time point on the contour.

init: Initial (zero) value of the integral.

seq: Quasi-random sequence generator.

N: The number of points to be taken from the quasi-random sequence.

Returns

Value of the integral.

QInchworm.qmc_integrate.qmc_inchworm_integral_root — Function

qmc_inchworm_integral_root(
    f,
    d_before::Int64,
    d_after::Int64,
    c::Keldysh.AbstractContour,
    t_i::Keldysh.BranchPoint,
    t_w::Keldysh.BranchPoint,
    t_f::Keldysh.BranchPoint;
    init,
    seq,
    N
)

Evaluate an inchworm-type contour-ordered integral of a function $f(\mathbf{t})$,

\[\int_{t_w}^{t_f} dt_1 \int_{t_w}^{t_1} dt_2 \ldots \int_{t_w}^{t_{d_{after}-1}} dt_{d_{after}} \int_{t_i}^{t_w} dt_{d_{after}+1} \ldots \int_{t_i}^{t_{d-2}} dt_{d-1} \int_{t_i}^{t_{d-1}} dt_d\ f(t_1, t_2, \ldots, t_d)\]

using the Sobol sequence for quasi-random sampling and a two-piece 'Root' transform. The total dimension of the domain d is the sum of the amounts of integration variables in the 'before' ($d_\text{before}$) and the after ($d_\text{after}$) components.

Parameters

f: Integrand.

d_before: Dimensionality of the before-$t_w$ component of the integration domain.

d_after: Dimensionality of the after-$t_w$ component of the integration domain.

c: Time contour to integrate over.

t_i: Starting time point on the contour.

t_w: 'Worm' time point on the contour separating the 'before' and 'after' parts.

t_f: Final time point on the contour.

init: Initial (zero) value of the integral.

seq: Quasi-random sequence generator.

N: The number of samples to be taken.

Returns

Value of the integral.

Miscellaneous

QInchworm.qmc_integrate.branch_direction — Constant

Dictionary mapping branches of the Keldysh contour to their unitary direction coefficients in the complex time plane.

QInchworm.qmc_integrate.contour_function_return_type — Function

contour_function_return_type(f::Function) -> Any

Detect the return type of a function applied to a vector of Keldysh.BranchPoint.