QInchworm.randomization

QInchworm.randomization.mean_std_from_randomization — Function

mean_std_from_randomization(
    f::Function,
    D::Int64,
    params::QInchworm.randomization.RandomizationParams;
    seq_type
) -> Tuple{Any, Any}

Estimate mean and standard deviation using randomized quasi Monte Carlo.

This function initializes a number of D-dimensional scrambled Sobol sequences using the set of parameters params and passes each of them to the function f. It returns the mean and the standard deviation of f's return values.

A different sequence type can be specified using the seq_type argument, provided its constructor has the same signature as ScrambledSobolSeq.