QInchworm.ppgf

QInchworm.ppgf.atomic_ppgf

QInchworm.ppgf.first_order_spgf

QInchworm.ppgf.operator_matrix_representation

QInchworm.ppgf.operator_product

QInchworm.ppgf.set_matsubara!

QInchworm.ppgf.set_ppgf_symmetric!

QInchworm.ppgf.atomic_ppgf — Method

Compute atomic pseudo-particle Green's function on the time grid for a time-independent problem defined by the EDCore instance.

QInchworm.ppgf.first_order_spgf — Method

Compute the first order pseudo-particle diagram contribution to the single-particle Green's function g_{o1, o2}(z, z')

QInchworm.ppgf.operator_matrix_representation — Method

Get matrix representation of operator expression in each sector

NB! Requires that the operator expression does not mix symmetry sectors

QInchworm.ppgf.operator_product — Method

operator_product(...)

Evaluate a product of vertices at different contour times z_i with the pseudo-particle Green's function sandwiched in between.

vertices is a contour-time ordered list of triples (z_i, c_i, o_i) were: z_i is a contour time, c_i is +1/-1 for creation/annihilation operator respectively, and o_i is a spin-orbital index

QInchworm.ppgf.set_matsubara! — Method

Set all time translation invariant values of the Matsubara branch

QInchworm.ppgf.set_ppgf_symmetric! — Method

Set real-time ppgf symmetry connected time pairs

NB! times has to be in the inching region with z2 ∈ backward_branch.