QInchworm.qmc_integrate

QInchworm.qmc_integrate.contour_function_return_type

QInchworm.qmc_integrate.exp_p_norm

QInchworm.qmc_integrate.make_exp_model_function

QInchworm.qmc_integrate.make_model_function

QInchworm.qmc_integrate.qmc_integral

QInchworm.qmc_integrate.qmc_integral_n_samples

QInchworm.qmc_integrate.contour_function_return_type — Method

Detect the return type of a function applied to a vector of branch points.

QInchworm.qmc_integrate.exp_p_norm — Method

Normalization of the model function p_d(u)

QInchworm.qmc_integrate.make_exp_model_function — Method

Make p_d(u) from the exponential h(v).

QInchworm.qmc_integrate.make_model_function — Method

Make the model function p_d(u) out of h_i(v).

QInchworm.qmc_integrate.qmc_integral — Function

Quasi Monte Carlo integration with warping.

`f`      Integrand.
`init`   Initial value of the integral.
`p`      Positive model function p_n(u).
`p_norm` Integral of p_n(u) over the u-domain.
`trans`  Transformation from x ∈ [0,1]^d onto the u-domain.
`seq`    Quasi-random sequence generator.
`N`      The number of points taken from the quasi-random sequence.

QInchworm.qmc_integrate.qmc_integral_n_samples — Function

Quasi Monte Carlo integration with warping.

This function takes a specified number of valid samples of the integrand.
`nothing` returned by the integrand does not count towards this number.

`f`         Integrand.
`init`      Initial value of the integral.
`p`         Positive model function p_n(u).
`p_norm`    Integral of p_n(u) over the u-domain.
`trans`     Transformation from x ∈ [0,1]^d onto the u-domain.
`seq`       Quasi-random sequence generator.
`N_samples` The number of taken samples.