QInchworm.utility

Interpolations.jl addon: Neumann boundary conditions for the cubic spline

QInchworm.utility.NeumannBC — Type

struct NeumannBC{GT<:Union{Nothing, Interpolations.GridType}, T<:Number} <: Interpolations.BoundaryCondition

Interpolations.jl addon: Implementation of the Neumann boundary conditions for the cubic spline.

Fields

gt::Union{Nothing, Interpolations.GridType}: Grid type

left_derivative::Number: Function derivative at the left boundary

right_derivative::Number: Function derivative at the right boundary

Interpolations.prefiltering_system — Method

prefiltering_system(
    _::Type{T},
    _::Type{TC},
    n::Int64,
    degree::Interpolations.Cubic{BC<:(QInchworm.utility.NeumannBC{Interpolations.OnGrid})}
) -> Tuple{WoodburyMatrices.Woodbury, Vector}

Compute the system used to prefilter cubic spline coefficients when using the Neumann boundary conditions.

Parameters

T, TC: Element types.

n: The number of rows in the data input.

degree: Interpolation degree information.

Returns

Woodbury matrix and the RHS of the computed system.

Quadratic spline that allows for incremental construction

QInchworm.utility.IncrementalSpline — Type

struct IncrementalSpline{KnotT<:Number, T<:Number}

Quadratic spline on an equidistant grid that allows for incremental construction.

Fields

knots::AbstractRange{KnotT} where KnotT<:Number: Locations of interpolation knots

data::Vector{T} where T<:Number: Values of the interpolated function at the knots

der_data::Vector{T} where T<:Number: Values of the interpolated function derivative at the knots

QInchworm.utility.extend! — Function

extend!(
    G_int::QInchworm.spline_gf.IncSplineImaginaryTimeGF,
    val
) -> Any

Extend the underlying IncrementalSpline objects stored in G_int with a value val.

extend!(
    spline::QInchworm.utility.IncrementalSpline,
    val
) -> Vector{T} where T<:Number

Add a segment to an incremental spline by fixing value val of the interpolated function at the next knot.

Sobol sequence including the initial point $(0, 0, \ldots)$

QInchworm.utility.SobolSeqWith0 — Type

mutable struct SobolSeqWith0{N} <: Sobol.AbstractSobolSeq{N}

Sobol low discrepancy sequence including the initial point $(0, 0, \ldots)$. The need for this modified version of Sobol.SobolSeq is discussed in JuliaMath/Sobol.jl#31.

Sobol.next! — Function

next!(s::QInchworm.utility.SobolSeqWith0) -> Vector{Float64}

Return the next point in the Sobol sequence s.

QInchworm.utility.arbitrary_skip! — Function

arbitrary_skip!(
    s::QInchworm.utility.SobolSeqWith0,
    n::Integer
)

Skip exactly n points in the Sobol sequence s.

Lazy matrix product

QInchworm.utility.LazyMatrixProduct — Type

mutable struct LazyMatrixProduct{T<:Number}

A matrix product of the form $A_N A_{N-1} \ldots A_1$.

Functions pushfirst!() and popfirst!() can be used to add and remove multipliers to/from the left of the product. The product is lazy in the sense that the actual multiplication takes place only when the eval!() function is called. The structure keeps track of previously evaluated partial products and reuses them upon successive calls to eval!().

Base.pushfirst! — Method

pushfirst!(
    lmp::QInchworm.utility.LazyMatrixProduct{T<:Number},
    A::Array{T<:Number, 2}
) -> Matrix{T} where T<:Number

Add a new matrix A to the left of the product lmp.

Base.popfirst! — Method

popfirst!(
    lmp::QInchworm.utility.LazyMatrixProduct{T<:Number}
) -> Int64
popfirst!(
    lmp::QInchworm.utility.LazyMatrixProduct{T<:Number},
    n::Int64
) -> Int64

Remove n matrices from the left of the product lmp. By default, n = 1.

QInchworm.utility.eval! — Method

eval!(
    lmp::QInchworm.utility.LazyMatrixProduct{T<:Number}
) -> Matrix{ScalarType} where ScalarType<:Number

Evaluate the matrix product lmp.

Serialization using IOBuffer

QInchworm.utility.iobuffer_serialize — Function

iobuffer_serialize(data)

Serialize data using an IOBuffer object.

QInchworm.utility.iobuffer_deserialize — Function

iobuffer_deserialize(data_raw)

Deserialize data using an IOBuffer object.

Range partitioning utilities

QInchworm.utility.split_count — Function

split_count(N::Integer, n::Integer) -> Any

Return a vector of n integers which are approximately equal and sum to N.

QInchworm.utility.range_from_chunks_and_idx — Function

range_from_chunks_and_idx(
    chunk_sizes::AbstractVector,
    idx::Integer
) -> Any

Given a list of chunk sizes, return the range that enumerates elements in the idx-th chunk.